dirty nut working version with piles for acoustic equation

+++ libsrc/DataInOut/Unverg/outUnverg.hh
@@ -27,13 +27,13 @@ public:
   virtual void WriteGrid(const Integer level);
 
   /// write information about the solution
-  virtual void WriteSolution(const Vector<Double> & sol, const Integer step, const Integer time);
+  virtual void WriteSolution(const Vector<Double> & sol, const Integer step, const Double time);
 
   /// write information about first derivatives of the solution
-  virtual void WriteFirstDerSolution(const Vector<Double> & sol, const Integer step, const Integer time);
+  virtual void WriteFirstDerSolution(const Vector<Double> & sol, const Integer step, const Double time);
 
   /// write information about second derivatives of the solution
-   virtual void WriteSecondDerSolution(const Vector<Double> & sol, const Integer step, const Integer time);
+   virtual void WriteSecondDerSolution(const Vector<Double> & sol, const Integer step, const Double time);
 
 private:
   ///
